// src/app/models/cliente.ts
export class cliente{
    idClientes: number=0;
    nombre: string='';
    pais: string='';
    estado: string='';
    ciudad: string='';
    CP: string='';
    colonia: string='';
    numero: number=0;
    calle: string='';
    email: string='';
    contrasenia : string='';
    eliminado: boolean=false;
    tipo: number=2;
    idCredencial: number=0;
    idUbicacion: number=0;
}